+++ code/ch09/optimized_memory_bound.py
@@ -108,7 +108,7 @@ class OptimizedMemoryBoundBenchmark(VerificationPayloadMixin, BaseBenchmark):
             total_flops=float(getattr(self, 'total_flops', getattr(self, 'N', 1024) * 2)),
             total_bytes=float(getattr(self, 'N', 1024) * 4 * 2),
             elapsed_ms=getattr(self, '_last_elapsed_ms', None),
-            precision="fp16",
+            precision="fp32",
         )
 
     def validate_result(self) -> Optional[str]:
@@ -126,5 +126,4 @@ class OptimizedMemoryBoundBenchmark(VerificationPayloadMixin, BaseBenchmark):
 def get_benchmark() -> BaseBenchmark:
     """Factory function for benchmark discovery."""
     return OptimizedMemoryBoundBenchmark()
-
